@@ -1,9 +1,21 @@
 source 'http://rubygems.org'
 
-gem 'rails', '3.0.4'
+# gem 'rails', '~> 3.1.0'
+
+# Bundle edge Rails instead:
+gem 'rails', :git => 'git://github.com/rails/rails.git', :branch => '3-1-stable'
+
 
 gem 'tiny_tds'
-gem 'activerecord-sqlserver-adapter'
+# gem 'activerecord-sqlserver-adapter'
+gem 'activerecord-sqlserver-adapter', :git => 'git://github.com/rails-sqlserver/activerecord-sqlserver-adapter.git'
+
+# Asset template engines
+gem 'sass-rails', "~> 3.1.0.rc"
+gem 'coffee-script'
+gem 'uglifier'
+
+gem 'jquery-rails'
 
 # Use unicorn as the web server
 # gem 'unicorn'
@@ -11,19 +23,10 @@ gem 'activerecord-sqlserver-adapter'
 # Deploy with Capistrano
 # gem 'capistrano'
 
-# To use debugger (ruby-debug for Ruby 1.8.7+, ruby-debug19 for Ruby 1.9.2+)
-# gem 'ruby-debug'
-# gem 'ruby-debug19'
+# To use debugger
+# gem 'ruby-debug19', :require => 'ruby-debug'
 
-# Bundle the extra gems:
-# gem 'bj'
-# gem 'nokogiri'
-# gem 'sqlite3-ruby', :require => 'sqlite3'
-# gem 'aws-s3', :require => 'aws/s3'
-
-# Bundle gems for the local environment. Make sure to
-# put test-only gems in this group so their generators
-# and rake tasks are available in development mode:
-# group :development, :test do
-#   gem 'webrat'
-# end
+group :test do
+  # Pretty printed test output
+  gem 'turn', :require => false
+end
